update dockerfile and readme; split demo into scripts and notebooks
