+++ docker/gpu/1.0/Dockerfile
@@ -8,4 +8,4 @@ WORKDIR /root
 RUN git clone https://github.com/UK-Digital-Heart-Project/4Dsurvival.git && cd 4Dsurvival && git fetch \
     && pip3 install --upgrade pip setuptools && python setup.py develop
 
-RUN pip3 install --upgrade keras==2.2.2 lifelines optunity matplotlib sklearn scipy pandas numpy
+RUN pip3 install --upgrade keras==2.2.2 lifelines==0.14.6 optunity matplotlib sklearn scipy pandas numpy
