+++ docker/gpu/1.1/Dockerfile
@@ -1,0 +1,3 @@
+FROM lisurui6/4dsurvival-gpu:1.0
+
+RUN pip3 uninstall lifelines && pip3 install lifelines==0.23.9
